refactored to package example html
