+++ jigs/test1/test1.example.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Test 1</title>
+</head>
+<body>
+    <section>
+        Test 1
+    </section>
+</body>
+</html>
