+++ jigs/test2/test2.example.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Test 2</title>
+</head>
+<body>
+<section>
+    Test 2
+</section>
+</body>
+</html>
