+++ jigs/test3/test3.example.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Test 3</title>
+</head>
+<body>
+<section>
+    Test 3
+</section>
+</body>
+</html>
